lib: consume also data side of the block
In this way, at the end of the read we really consumed the entire
block. Otherwise the block still pending till the next control read
and this may cause problem with the ZIO circular buffer.

This is not an hack, this is how ZIO expects user-space read data.
The hack was the previous situation

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>