kernel: add WR manual offset for old calibration data
Apparently the White-Rabbit calibration offset is wrong for the FMC-TDC
mezzanines calibrated before 2018-01-01 (in the future, we are still fixing
the issue). this patch allows to overwrite that value.

It adds also the carrier components. The total WR offset depends on the carrier
as well. The calibration is "always" done with a SPEC, so in principle you
should add/subtract a carrier offset only if you are using a board
different than SPEC

Note
The problem is not easy to spot and there are many FMC-TDC mezzanine produced
and already calibrated. The good thing to do is to replace the value on the
eeprom for all of them, but we cannot. So, we need this hack

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>