drv: use ZIO buffer to store timestamps
With DMA transfers is sensible to store all our data directly in the ZIO
buffer. ZIO timestamp support is not good today, so we need to act like if we
are acquiring generic data (why not?) where our sample is a timestamp.

Here we just do a quick and dirty thing to move the same timestamp from
the ctrl and from data

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>