*: set timeout coalescing dynamically
The driver exports a sysfs attrubte to change the coalescing timeout. A new
library funciton can be used to set/get this value; and the test program allows
the configuration before an acquisition.

Keep in mind that, for FIFOs the configuration is global (users are warned on
dmesg) while for DMA it is per-channel. I though it was better to have a common
interface for both. Users are warnd in the documentation as well. In the long
run, perhaps also the FIFO can have its own register.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>