ut: operational tests need White-Rabbit
In order for the test to work, both TDC and FD must share the same time.
It is possible to achive this with the local clock on both cards, but this
is not precise, I think that at best we can have a resolution of some
milliseconds.

I prefer to do a serious test.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>