sw:drv: DMA bugfix
The kernel API sg_alloc_table_from_pages() takes a start offset. The
problem is that we may use vmalloc() which have a start offset, but it
could have an offset on each page. Not considering this offset leads
to DMA transfers which are partially performed on the wrong buffer

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>