software: kernel: fmc-tdc: Update API for '(*read)' in 'struct memory_ops'
In struct memory_ops, the function-pointer (*read) is declared as
	u32 (*)(void *)

In its use, its object is given the address of function 'ioread32' (or
'ioread32be') which used to be of the same type.

Since linux-v5.8, their protype has been changed to
	u32 (*)(const void *)
at:
commit 8f28ca6bd8211214faf717677bbffe375c2a6072 ("iomap: constify ioreadX()
iomem argument (as in generic implementation)")

Hence, update the declaration of (*read) as well.

Signed-off-by: Vaibhav Gupta <vaibhav.gupta@cern.ch>