software: kernel: fmc-spec-core: Add 'linux/platform_device.h'
The 'struct platform_device_id' is defined in 'linux/mod_devicetable.h';
this header was #include in 'linux/platform_device.h'. Hence, using the
latter header-file was enough. But in linux-v4.18 at:

	commit ac3167257b9fe16c9426c2087ead1c9f1b0992b1 ("headers: separate
	linux/mod_devicetable.h from linux/platform_device.h")

the former was no more included in the latter. Thus, we need to do it
explicitly.

Since this change makes no difference if this driver is compiled against
Kernel version less than v4.18, we do not need any conditional
directives.

Signed-off-by: Vaibhav Gupta <vaibhav.gupta@cern.ch>