software: kernel: second parameter of ft_temperature_get() is not required
The second paramter of ft_temperature_get() is already a member of the
of the first one. Thus, the value can be stored in it within the
function itself. Aand, it doesn't changes its return behaviour.

Also, this makes the function symmetric, in terms of declaration, to
those of fmc-adc and fine-delay.

Signed-off-by: Vaibhav Gupta <vaibhav.gupta@cern.ch>