software: kernel: Use HWMON API to expose Temperature
Information from sensors like temperature, which are useful for
user-space, can be exposed via sysfs.

Earlier, the driver developers had to take care of defining the attributes,
initiating them, and perform other related jobs.

But since a long time, the HWMON framework in Linux has developed
significantly. And, utilizing it will reduce lot of maintainence work
for the developers.

All the sysfs related part is handled by the framework which also allows
user-space tools, like lm-sensors, to provide the data to the user in a
simplified manner.

Signed-off-by: Vaibhav Gupta <vaibhav.gupta@cern.ch>